feat(op): implement operations for tensor manipulation including read, write, fill, narrow, and to_host
